Reject cross-program final fn state writes
